[layers/common] Pass dtype to make_array_from_callback in general_device_put
The multi-host branch of general_device_put builds the global array with

    jax.make_array_from_callback(t.shape, sharding, lambda index: t[index])

and lets jax infer the dtype from what the callback returns. jax only
invokes that callback for shards this process addresses, so on a process
that owns none it has nothing to infer from and raises:

    ValueError: If the Array has no addressable shards, `dtype` must be
    provided via the `dtype` argument to `jax.make_array_from_callback`.

Every host holds a shard of a purely tensor-parallel array, which is why
this has not been hit before. It appears as soon as an axis is not spread
across all hosts -- expert parallelism being the ordinary case, where a
host holds no shard of the experts it was not assigned. Loading
Qwen3.5-397B-A17B on a 32-chip, 8-host Ray rollout at
tensor_parallelism=2 x expert_parallelism=8 fails this way in
tpu_runner.load_model, after the weights have already been placed
(hbm=(54.31, 95.73) per chip).

The source tensor's dtype is known, so pass it. No behaviour change when
a shard is addressable: jax checks the inferred dtype against the
argument and they agree.

(cherry picked from commit cd2acb0beec674db4cecac8167b2f18801d02148)

+++ tpu_inference/layers/common/utils.py
@@ -156,8 +156,18 @@ def general_device_put(tensor: jax.Array,
         # `t[i]` needs to be operated in the same mesh as `t`, which is provided as
         # `source_mesh`.
         with ctx:
+            # `dtype` is passed explicitly because jax infers it from the
+            # callback's return value, and the callback is never invoked on a
+            # process that owns no addressable shard of `sharding`:
+            #   ValueError: If the Array has no addressable shards, `dtype` must
+            #   be provided via the `dtype` argument to
+            #   `jax.make_array_from_callback`.
+            # Every host holds a shard of a purely tensor-parallel array, so
+            # this only shows up once some axis is not spread over all hosts --
+            # expert parallelism being the common case, where a host holds no
+            # shard of the experts it was not assigned.
             global_array = jax.make_array_from_callback(
-                t.shape, sharding, lambda index: t[index])
+                t.shape, sharding, lambda index: t[index], dtype=t.dtype)
         if layout is not None:
             dst_mesh = sharding.mesh
             with jax.set_mesh(dst_mesh):
